tests/wpt/tests/css/css-transitions/all-with-discrete.tentative.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/w3c/csswg-drafts/issues/4441">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-animations/support/testcommon.js"></script>

<div class=testcase id=discrete>Allow discrete</div>
<div class=testcase id=normal>No discrete</div>

<style>
.testcase {
  float: left;
  width: 100px;
  height: 100px;
  transition: all 1s;
}
#discrete {
  transition-behavior: allow-discrete;
}
.animated {
  float: right;
}
</style>

<script>
promise_test(async () => {
  await waitForAnimationFrames(2);
  // Regression test for https://crbug.com/1518561
  normal.classList.add('animated');
  const normalAnims = normal.getAnimations();
  assert_equals(normalAnims.length, 0,
                "Started a discrete property transition on a normal element");

  discrete.classList.add('animated');
  const anims = discrete.getAnimations();
  assert_equals(anims.length, 1,
                "Did not start a discrete-property transition");
  assert_equals(anims[0].transitionProperty, 'float');
}, 'transition:all with transition-behavior:allow-discrete should animate discrete properties.');
</script>
